added memory flags in dockers + modified model and orchestration + added more debug and build features

+++ src/main/java/model/elastic/Event_Stub.java
@@ -4,6 +4,6 @@ import model.TestStub;
 
 public interface Event_Stub extends TestStub {
 	DetectedObject_Stub get_detected_object();
-	long get_timestamp();
+	int get_timestamp();
 	void delete();
 }
